# openapi.json
{
  "components": {
    "schemas": {
...
      "ExperienceLevel": {
        "enum": [
          "entry",
          "mid",
          "senior",
          "lead",
          "executive"
        ],
        "title": "ExperienceLevel",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobSearchResponse": {
        "description": "Schema for paginated job search results.",
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Results",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "page_size",
          "results"
        ],
        "title": "JobSearchResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "draft",
          "open",
          "on_hold",
          "filled",
          "closed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobType": {
        "enum": [
          "full_time",
          "part_time",
          "contract",
          "temporary",
          "internship",
          "freelance"
        ],
        "title": "JobType",
        "type": "string"
      },
      "LocationType": {
        "enum": [
          "on_site",
          "remote",
          "hybrid"
        ],
        "title": "LocationType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/interviews/job/{job_id}": {
      "get": {
        "description": "Get all interviews for a specific job.",
        "operationId": "get_interviews_for_job_api_interviews_job__job_id__get",
        "parameters": [
          {
            "description": "The job ID to get interviews for",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "description": "The job ID to get interviews for",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Interview"
                  },
                  "title": "Response Get Interviews For Job Api Interviews Job  Job Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Interviews For Job",
        "tags": [
          "interviews",
          "interviews"
        ]
      }
    },
...
    "/api/jobs": {
      "get": {
        "description": "Get all jobs, optionally filtered by status.\n\nArgs:\n    status: Optional filter for job status (e.g., 'active', 'open')\n    page: Page number (1-based)\n    page_size: Number of items per page (max 100)\n    sort_by: Field to sort by\n    sort_order: Sort order ('asc' or 'desc')\n    \nReturns:\n    List of jobs with pagination info",
        "operationId": "get_jobs_api_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "created_at",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "default": "desc",
              "title": "Sort Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/": {
      "get": {
        "description": "Search for jobs with various filters.",
        "operationId": "search_jobs_api_jobs__get",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "in": "query",
            "name": "department",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department"
            }
          },
          {
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "in": "query",
            "name": "location_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LocationType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Type"
            }
          },
          {
            "in": "query",
            "name": "job_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Type"
            }
          },
          {
            "in": "query",
            "name": "experience_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ExperienceLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experience Level"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "min_salary",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Salary"
            }
          },
          {
            "in": "query",
            "name": "max_salary",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Salary"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "created_at",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "default": "desc",
              "title": "Sort Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Create a new job posting.\nNOTE: Always POST to /jobs/ (with trailing slash) to avoid redirect issues.",
        "operationId": "create_job_api_jobs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/applications/{candidate_id}": {
      "get": {
        "description": "Get all applications for a candidate.",
        "operationId": "get_candidate_applications_api_jobs_applications__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CandidateApplicationSummary"
                  },
                  "title": "Response Get Candidate Applications Api Jobs Applications  Candidate Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate Applications",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/track-view": {
      "post": {
        "description": "Track when a job is viewed (for analytics).",
        "operationId": "track_job_view_api_jobs__job_id__track_view_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackViewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Job View",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}